+++ R/capFirst.R
@@ -16,22 +16,27 @@
 #' capFirst("Derek Ogle")
 #' capFirst("derek ogle")
 #' capFirst("derek")
-#' # apply to all elements in a vector
-#' vec <- c("Derek Ogle","derek ogle","Derek ogle","derek Ogle","DEREK OGLE")
-#' apply(matrix(vec),1,capFirst)
 #'
 #' ## Capitalize first letter of only the first words
 #' capFirst("Derek Ogle",words="first")
 #' capFirst("derek ogle",words="first")
 #' capFirst("derek",words="first")
-#' # apply to all elements in a vector
-#' apply(matrix(vec),1,capFirst,words="first")
+
+#' ## apply to all elements in a vector
+#' vec <- c("Derek Ogle","derek ogle","Derek ogle","derek Ogle","DEREK OGLE")
+#' capFirst(vec)
+#' capFirst(vec,words="first")
 #'
 #' @export
-capFirst<- function(x,words=c("all","first")) {
+capFirst <- function(x,words=c("all","first")) {
+  if (length(x)==1) iCapFirst(x,words)
+  else apply(matrix(x),MARGIN=1,FUN=iCapFirst,words=words)
+}
+
+iCapFirst<- function(x,words=c("all","first")) {
   # See whether all or just the first word should have the first letter capitalized
   words <- match.arg(words)
-  # convet entire string to lower case ...
+  # convert entire string to lower case ...
   x <- tolower(x)
   # then split on space if more than one word
   s <- strsplit(x, " ")[[1]]
@@ -47,3 +52,4 @@ capFirst<- function(x,words=c("all","first")) {
     paste(s1,substring(s,2),sep="",collapse=" ")
   }
 }
+