+++ R/psdDataPrep.R
@@ -1,10 +1,10 @@
-#' @title Add five-cell length categories specific to a species to an entire data frame.
+#' @title Add Gabelhouse lengths for each species in an entire data frame.
 #'
-#' @description Adds the five-cell length categories specific to a species to all individuals in an entire data frame.  The concept is to save the user from having to split each species into a separate data frame, then creating the required length categorization variable, and then binding the separate data frames back together.
+#' @description Adds the Gabelhouse lengths specific to a species to all individuals in an entire data frame.  The concept is to save the user from having to split each species into a separate data frame, then create the required length categorization variable, and then bind the separate data frames back together.
 #'
-#' @details This function requires that the data frame provided in \code{data} contains one column of length measurements and one column of species names.  All lengths must be in the same units.  Any species name that is not recognized as being a species with known five-cell length category values (and stored in \code{data(PSDlit)}) will be considered as an \dQuote{other} type of fish (see below for how these fish can be handled).  Species that are known to have five-cell length categories can be found with \code{\link{psdVal}()} (i.e., without any arguments).  Spelling and capitalization have to be exactly the same as in \code{data(PSDlit)}) (i.e., generally, capitalized first letter for each word in the common name; see \code{\link{recodeSpecies}} for one method to change species names before using this function).
+#' @details The data frame provided in \code{data} must contain one column of length measurements and one column of species names.  All lengths must be in the same units.  Any species name that is not recognized as being a species with known Gabelhouse lengths will be considered as an \dQuote{other} type of fish (see below for how these fish can be handled).  Species that are known to have Gabelhouse length can be found with \code{\link{psdVal}()} (i.e., without any arguments).  Spelling and capitalization of species names have to be exactly (except for differences in capitalization) the same as in \code{data(PSDlit)}.  See \code{\link{recodeSpecies}} for one method to change species names before using this function).
 #'
-#' A new variable will be created in the returned data frame that contains the five-cell length categories that correspond to the given lengths.  If \code{use.catnames=TRUE} (the default) then the \dQuote{levels} of the new  variable will be the word names of the five-cell categories -- i.e., \dQuote{stock}, \dQuote{Quality}, etc., with the possibility that \dQuote{sub-stock} length fish will be labeled with \dQuote{zero}.  If \code{use.catnames=FALSE} then the levels of the new variable will be the numeric value representing the minimum value in each five-cell category, again with the possibility that \dQuote{sub-stock} fish will be given a value of zero.
+#' A new variable will be created in the returned data frame that contains the Gabelhouse lengths that correspond to the observed lengths.  If \code{use.names=TRUE} (the default) then the \dQuote{levels} of the new variable will be the names of the Gabelhouse length categories -- i.e., \dQuote{stock}, \dQuote{Quality}, etc., with the possibility that \dQuote{sub-stock} length fish will be labeled with \dQuote{zero}.  If \code{use.names=FALSE} then the levels of the new variable will be the numeric value representing each Gabelhouse length, again with the possibility that \dQuote{sub-stock} fish will be given a value of zero.
 #'
 #' The default behavior is for \dQuote{sub-stock} fish and \dQuote{other} fish (see above for the definition of an \dQuote{Other} fish) to be included in the resulting data frame, but with a \code{0} or \code{zero} in the new length categorization variable for the sub-stock fish and a \code{NA} in the new length categorization variable for the \dQuote{other} fish.  The \dQuote{sub-stock} fish can be excluded from the final data frame with \code{remove.substock=TRUE} and the \dQuote{other} fish can be excluded with \code{remove.other=TRUE}.
 #'
@@ -12,17 +12,17 @@
 
 #' @param formula A formula of the form \code{length~species} where \dQuote{length} generically represents a variable that contains length measurements and \dQuote{species} generically represents a variable that contains species names.  Note that this formula can only contain two variables and must have the length variable on the left-hand-side and the species variable on the right-hand-side.
 #' @param data A data.frame that minimally contains the length measurements and species names given in the variables in the \code{formula}.
-#' @param psdname A string that contains the name for the new length category variable (defaults to \dQuote{PSDcat}).
-#' @param units A string that indicates the type of units used for the length measurements.  Choices are \code{mm} for millimeters (DEFAULT), \code{cm} for centimeters, and \code{in} for inches.
-#' @param remove.substock A logical that indicates whether to exclude individual fish with length measurements that are less than the \dQuote{stock} size for that species.  See details.
-#' @param remove.other A logical that indicates whether to exclude all individuals of a fish species for which the five-cell length categories do not exist in \code{data(PSDlit)}.  See details.
-#' @param use.catnames A logical that indicates whether the categories returned in \code{psdname} are numeric (\code{=FALSE}) or word (\code{=TRUE}; default) representations of the five-cell categories.  See details.
+#' @param vname A string that contains the name for the new length category variable (defaults to \dQuote{PSDcat}).
+#' @param units A string that indicates the type of units used for the lengths.  Choices are \code{mm} for millimeters (DEFAULT), \code{cm} for centimeters, and \code{in} for inches.
+#' @param remove.substock A logical that indicates whether to exclude individual fish with lengths that are less than the \dQuote{stock} size for that species.  See details.
+#' @param remove.other A logical that indicates whether to exclude all individuals of a fish species for which the Gabelhous lengths do not exist in \code{data(PSDlit)}.  See details.
+#' @param use.names A logical that indicates whether the categories returned in \code{vname} are numeric (\code{=FALSE}) or word (\code{=TRUE}; default) representations of the Gabelhouse lengths.  See details.
 #'
-#' @return The original data frame in \code{data} but with a new variable \code{psdname} appended that contains the PSD-related length categories.
+#' @return The original data frame in \code{data} but with a new variable \code{vname} appended that contains the PSD-related length categories.
 #' 
 #' @author Derek H. Ogle, \email{dogle@@northland.edu}
 #' 
-#' @seealso \code{\link{psdVal}}, \code{\link{psdCalc}}, \code{\link{psdPlot}}, \code{\link{PSDlit}}, \code{\link{tictactoe}}, \code{\link{tictactoeAdd}}, \code{\link{lencat}}, \code{\link{rcumsum}}, \code{\link{recodeSpecies}}, and \code{\link{wrDataPrep}}.
+#' @seealso \code{\link{psdVal}}, \code{\link{psdCalc}}, \code{\link{psdPlot}}, \code{\link{PSDlit}}, \code{\link{tictactoe}}, \code{\link{lencat}}, \code{\link{rcumsum}}, \code{\link{recodeSpecies}}, and \code{\link{wrDataPrep}}.
 #'
 #' @section fishR vignette: none yet.
 #' 
@@ -54,7 +54,7 @@
 #' # Example where species without PSD values and sub-stock fish are retained (the defaults)
 #' d3 <- psdDataPrep(tl~species1,data=d2,units="mm")
 #' str(d3)
-#' with(d3,table(species1,PSDcat))
+#' xtabs(~species1+PSDcat,data=d3)
 #'
 #' # Example where species without known PSD values are removed, but sub-stock fish are not
 #' d4 <- psdDataPrep(tl~species1,data=d2,units="mm",remove.other=TRUE)
@@ -68,60 +68,56 @@
 #' with(d5,table(species1)) 
 #'
 #' # Example where category names are not used
-#' d6 <- psdDataPrep(tl~species1,data=d2,units="mm",use.catnames=FALSE)
+#' d6 <- psdDataPrep(tl~species1,data=d2,units="mm",use.names=FALSE)
 #' str(d6)
 #' with(d6,table(species1,PSDcat))
 #'
 #' @export
-psdDataPrep <- function(formula,data,psdname="PSDcat",
+psdDataPrep <- function(formula,data,vname="PSDcat",
                         units=c("mm","cm","in"),
                         remove.substock=FALSE,remove.other=FALSE,
-                        use.catnames=TRUE) {
+                        use.names=TRUE) {
+  ## Some checks
   units <- match.arg(units)
+  # get variable names and check data type
   cl <- iGetVarFromFormula(formula,data,expNumVars=2)
   cspp <- cl[2]
+  if (is.numeric(data[,cspp])) stop("Variable on RHS of 'formula' is numeric (and, thus, not species names).",call.=FALSE)
   cl <- cl[1]
-  incl.zero <- !remove.substock
-  # get the PSD literature values data frame
+  if (!is.numeric(data[,cl])) stop("Variable on LHS of 'formula' is non-numeric (and, thus, not fish lengths).",call.=FALSE)
+  ## get the PSD literature values data frame
   PSDlit <- get(data("PSDlit", envir = environment()), envir = environment())
-  # what species exist in the data frame
+  ## if supposed to remove substock fish then don't include zero in PSD breaks
+  incl.zero <- !remove.substock
+  ## what species exist in the data frame
   specs <- levels(data[,cspp])
   # which species in the data frame are and are not found in the PSDlit list
   areLogical <- specs %in% levels(PSDlit$species)
   noPSD <- specs[which(!areLogical)]
-  if (length(noPSD)>0 & remove.other) warning("The following species do not have known Gabelhouse five-cell length categories:\n ",paste(noPSD,collapse=", "),call.=FALSE)
+  if (length(noPSD)>0) warning("These species do not have known Gabelhouse lengths:\n ",paste(noPSD,collapse=", "),call.=FALSE)
   specs <- specs[which(areLogical)]
   
-  # initiate a blank data frame that has same columns as old data frame
-  ndf <- data[-c(1:nrow(data)),]
   # cycle through each species where PSD values are known
-  for (i in 1:length(specs)) {
+  for (i in specs) {
     # get the Gabelhouse length categories
-    glhse <- psdVal(specs[i],units=units,incl.zero=incl.zero)
+    glhse <- psdVal(i,units=units,incl.zero=incl.zero)
     # isolate the current species
-    idf <- data[data[,cspp]==specs[i],]
+    idf <- data[data[,cspp]==i,]
     # if incl.zero not used then eliminate all fish below stock size
     if (!incl.zero) idf <- idf[idf[,cl]>=glhse[1],]
     # add the Gabelhouse length categories to the data frame
-    idf <- lencat(as.formula(paste("~",cl,sep="")),data=idf,breaks=glhse,vname=psdname,use.catnames=use.catnames)
-    # if asked, change numerical category names to stock, quality, etc. names
-    if (use.catnames) {
-      idf[,psdname] <- factor(idf[,psdname])
-      # must use match so that names of levels line up with numerical values
-      levels(idf[,psdname]) <- names(glhse)[match(levels(idf[,psdname]),glhse)]    
-    }
+    idf <- lencat(as.formula(paste("~",cl,sep="")),data=idf,breaks=glhse,vname=vname,use.names=use.names)
     # bind current species to the new data frame being created
-    ndf <- rbind(ndf,idf)
+    if (i==specs[1]) ndf <- idf
+      else ndf <- rbind(ndf,idf)
   }
   # add back in "other" fish if asked not to remove
   if (!remove.other & length(noPSD)>0) {
     odf <- data[data[,cspp] %in% noPSD,]
-    odf[,psdname] <- NA
+    odf[,vname] <- NA
     ndf <- rbind(ndf,odf)
   }
-  # make sure the new length categories and species name variables are factored
-  if (use.catnames) ndf[,psdname] <- factor(ndf[,psdname],levels=names(glhse))
-    else ndf[,psdname] <- factor(ndf[,psdname])
+  # factor the species variable so that it will remove other species names if they are not include
   ndf[,cspp] <- factor(ndf[,cspp])
   # return the new data frame
   ndf