+++ R/psdVal.R
@@ -1,26 +1,26 @@
-#' @title Finds Gabelhouse five-cell length categories (for PSD) for a particular species.
+#' @title Finds Gabelhouse lengths (for PSD calculations) for a species.
 #'
-#' @description Returns a vector with the Gabelhouse five-cell length categories for a particular species.
+#' @description Returns a vector with the five Gabelhouse lengths for a chosen species.
 #'
-#' @details Finds the Gabelhouse five-cell length categories for the given species names in the \code{data(PSDlit)} data frame.  The species names must be spelled exactly like they appear in the \code{data(PSDlit)} data frame (type \code{psdVal()} to see the list of species and how they are spelled).
+#' @details Finds the Gabelhouse lengths from the \code{data(PSDlit)} data frame for the species given in \code{species}.  The species name must be spelled exactly (within capitalization differences) as it appears in \code{data(PSDlit)} (type \code{psdVal()} to see the list of species and how they are spelled).
 #'
-#' If \code{incl.zero=TRUE} then a zero will be included in the first postion of the returned vector and the five-cell length categories will follow.  This is useful when computing PSD values as a sample of fish may have individuals smaller than the stock length.
+#' A zero is included in the first postion of the returned vector if \code{incl.zero=TRUE}.  This is useful when computing PSD values with a data.frame that contains fish smaller than the stock length.
 #'
-#' The \code{addLens} argument can be used to add additional lengths to the vector of length categories returned.  Names for these lengths can be included in the \code{addNames} argument.  If \code{addNames} is non-NULL then it must of the same length as \code{addLens}.  If \code{addLens} is non-NULL but \code{addNames} is NULL then the default names will be the same as the lengths in \code{addLens}.  The \code{addLens} argument is useful for calculating PSD values that are different from the usual five-cell categories.
+#' Additional lengths can be added to the returned vector with \code{addLens}.  Names for these lengths can be included in \code{addNames}.  If \code{addNames} is non-NULL then it must be of the same length as \code{addLens}.  If \code{addLens} is non-NULL but \code{addNames} is NULL then the default names will be the same as the lengths in \code{addLens}.  The \code{addLens} argument is useful for calculating PSD values that are different from the Gabelhouse lengths.
 #'
 #' @aliases psdVal
 #'
-#' @param species A string that contains the species name to find coefficients for.
-#' @param units A string that indicates the type of units used for the length measurements.  Choices are \code{mm} for millimeters (DEFAULT), \code{cm} for centimeters, and \code{in} for inches.
-#' @param incl.zero A logical that indicates whether to include a zero in the first position of the returned vector (DEFAULT) or not.  See details.
+#' @param species A string that contains the species name for which to find Gabelhous lengths.  See details.
+#' @param units A string that indicates the units for the returned lengths.  Choices are \code{mm} for millimeters (DEFAULT), \code{cm} for centimeters, and \code{in} for inches.
+#' @param incl.zero A logical that indicates if a zero is included in the first position of the returned vector (DEFAULT) or not.  See details.
 #' @param addLens A numeric vector that contains minimum length definitions for additional categories.  See details.
 #' @param addNames A string vector that contains names for the additional length categories added with \code{addLens}.  See details.
 #'
-#' @return A vector of the five-cell length categories for the given species.
+#' @return A vector of minimum values for length categories for the chosen species.
 #'
 #' @author Derek H. Ogle, \email{dogle@@northland.edu}
 #'
-#' @seealso \code{\link{psdCalc}}, \code{\link{psdPlot}}, \code{\link{psdDataPrep}}, \code{\link{PSDlit}}, \code{\link{tictactoe}}, \code{\link{tictactoeAdd}}, \code{\link{lencat}}, and \code{\link{rcumsum}}.
+#' @seealso \code{\link{psdCalc}}, \code{\link{psdPlot}}, \code{\link{psdDataPrep}}, \code{\link{PSDlit}}, \code{\link{tictactoe}}, \code{\link{lencat}}, and \code{\link{rcumsum}}.
 #'
 #' @section fishR vignette: \url{https://sites.google.com/site/fishrfiles/gnrl/SizeStructure.pdf}
 #'
@@ -57,10 +57,8 @@ psdVal <- function(species="List",units=c("mm","cm","in"),incl.zero=TRUE,
   # load RSDlit data frame into this function's environment
   # the data/get combination are used to avoid the "no global binding" note at CHECK
   PSDlit <- get(data("PSDlit", envir = environment()), envir = environment())
-  # check on PSDLit and species names
-  OK <- iPSDLitCheck(PSDlit,species <- capFirst(species))
   # continue if species name is correct
-  if (!is.null(OK)) {
+  if (iPSDLitCheck(PSDlit,species <- capFirst(species))) {
     # identify columns based on units
     ifelse(units=="in",cols <- 2:6,cols <- 7:11)
     # get the length categories
@@ -86,12 +84,15 @@ psdVal <- function(species="List",units=c("mm","cm","in"),incl.zero=TRUE,
 }
 
 
-
+# ============================================================
+# Internal function to check species name relative to the
+#   data.frame in 'data' (usually PSDLitCheck)
+# ============================================================
 iPSDLitCheck <- function(data,species) {
   OK <- FALSE
-  if (species=="List") iListSpecies(data)
-  else if (!any(levels(data$species)==species)) {
-    stop("The five-cell categories do not exist for your choice of species.\n  Please type psdVal() for a list of available species.\n\n",call.=FALSE)
-  }
-  else OK <- TRUE
+  if (length(species)!=1) stop("'species' can have only one name.",call.=FALSE)
+  else if (species=="List") iListSpecies(data)
+       else if (!any(levels(data$species)==species)) stop("The Gabelhouse lengths do not exist for ",species,".\n  Type psdVal() for a list of available species.\n\n",call.=FALSE)
+            else OK <- TRUE
+  OK
 }