+++ R/tictactoe.R
@@ -1,8 +1,8 @@
-#' @title Constructs the base tic-tac-toe graph for presenting predator-prey PSD values.
+#' @title Construct a base tic-tac-toe plot for presenting predator-prey PSD values.
 #'
-#' @description Constructs the base tic-tac-toe graph for presenting predator-prey PSD values.  Predator prey values are added with \code{\link{tictactoeAdd}}.
+#' @description Construct a base tic-tac-toe plot for presenting predator-prey PSD values.  Predator-prey PSD values are added with \code{plotCI} from \pkg{plotrix}.
 #'
-#' @details This function simply creates the base tic-tac-toe graph.  Observed values, with confidence intervals, can be added to this graph with \code{\link{tictactoeAdd}}.
+#' @details This function simply creates a base tic-tac-toe plot.  Observed values, with confidence intervals, are added to this plot with \code{plotCI} from \pkg{plotrix}.
 #'
 #' @param predbal A vector of length 2 that contains the range of values to be considered \dQuote{in balance} for the predator.
 #' @param preybal A vector of length 2 that contains the range of values to be considered \dQuote{in balance} for the prey.
@@ -10,40 +10,53 @@
 #' @param ylab A string representing a label for the y-axis.
 #' @param bal.col A string designating a color to which the \dQuote{in balance} regions should be shaded.
 #' @param bal.trans A numeric (decimal) that indicates the level of transparency for marking the \dQuote{in balance} regions.
-#' @param bnd.col A string designating a color for the boundaries of the \dQuote{in balance} regions.
-#' @param bnd.lwd A numeric designating the line width for the boundaries of the \dQuote{in balance} region.
-#' @param bnd.lty A numeric designating the line type for the boundaries of the \dQuote{in balance} regions.
+#' @param bnd.col A string that indicates a color for the boundaries of the \dQuote{in balance} regions.
+#' @param bnd.lwd A numeric that indicates the line width for the boundaries of the \dQuote{in balance} region.
+#' @param bnd.lty A numeric that indicates the line type for the boundaries of the \dQuote{in balance} regions.
 #'
 #' @return None.  However, a graphic is produced.
 #'
 #' @author Derek H. Ogle, \email{dogle@@northland.edu}
 #'
-#' @seealso \code{\link{tictactoeAdd}}, \code{\link{psdVal}}, \code{\link{psdCalc}}, \code{\link{psdPlot}}, \code{\link{psdDataPrep}}, \code{\link{PSDlit}}.
+#' @seealso \code{\link{psdVal}}, \code{\link{psdCalc}}.
 #'
 #' @section fishR vignette: \url{https://sites.google.com/site/fishrfiles/gnrl/SizeStructure.pdf}
 #'
 #' @keywords hplot
 #'
 #' @examples
-#' op <- par(mfrow=c(2,2))
-#'
-#' ## Simple plot with one point on it
+#' ## Create hypothetical data for plotting one point .. similar to what might come from psdCalc()
+#' prey <- c(45.4,30.2,56.8)
+#' pred <- c(24.5,10.2,36.7)
+#' names(prey) <- names(pred) <- c("Estimate","95% LCI","95% UCI")
+#' prey
+#' pred
+#' 
 #' tictactoe()
-#' tictactoeAdd(c(50,80),c(25,45))
-#' ## Changed predator balance lines, color of point
-#' tictactoe(predbal=c(20,50))
-#' tictactoeAdd(c(50,80),c(25,45),pt.col="blue")
-#' ## Add two points to plot
+#' if (require(plotrix)) {
+#'   plotCI(prey[1],pred[1],li=prey[2],ui=prey[3],err="x",pch=16,add=TRUE)
+#'   plotCI(prey[1],pred[1],li=pred[2],ui=pred[3],err="y",pch=16,add=TRUE) 
+#' }
+#' 
+#' ## Create hypothetical data for plotting three points .. similar to what might come from psdCalc()
+#' prey <- rbind(c(45.4,30.2,56.8),
+#'               c(68.2,56.7,79.4),
+#'               c(17.1, 9.5,26.3))
+#' pred <- rbind(c(24.5,10.2,36.7),
+#'               c(14.2, 7.1,21.3),
+#'               c(16.3, 8.2,24.4))
+#' colnames(prey) <- colnames(pred) <- c("Estimate","95% LCI","95% UCI")
+#' prey
+#' pred
+#' 
 #' tictactoe()
-#' tictactoeAdd(c(50,80),c(25,45),pt.col="blue")
-#' tictactoeAdd(c(30,80),c(35,45),pt.col="green")
-#' ## Add two points to plot, and label points
-#' tictactoe()
-#' tictactoeAdd(c(50,80),c(25,45),pt.col="blue",label="2007",lbl.cex=1.5)
-#' tictactoeAdd(c(30,80),c(35,45),pt.col="green",label="2009",lbl.pos="SW",lbl.cex=1.5)
-#'
-#' par(op)
-#'
+#' if (require(plotrix)) {
+#'   plotCI(prey[,1],pred[,1],li=prey[,2],ui=prey[,3],err="x",pch=16,add=TRUE)
+#'   plotCI(prey[,1],pred[,1],li=pred[,2],ui=pred[,3],err="y",pch=16,add=TRUE)
+#' }
+#' lines(prey[,1],pred[,1])
+#' text(prey[,1],pred[,1],labels=c(2010,2011,2012),adj=c(-0.5,-0.5))
+#' 
 #' @export
 tictactoe <- function(predbal=c(30,70),preybal=c(30,70),xlab="Predator PSD",ylab="Prey PSD",
                       bal.col="wheat",bal.trans=0.3,bnd.col="red",bnd.lwd=2,bnd.lty=2) {